+++ tests/authorization/test_api.py
@@ -4,22 +4,22 @@ from rest_framework.reverse import reverse
 def test_api():
     """Test API routes"""
     assert (
-        reverse("authorization:permission-set-v1-detail", kwargs=dict(pk=123))
-        == "/authorization/v1/permission-set/123/"
+        reverse("authorization:permission-set-v2-detail", kwargs=dict(pk=123))
+        == "/authorization/v2/permission-set/123/"
     )
     assert (
-        reverse("authorization:add-user-v1", kwargs=dict(pk=123))
-        == "/authorization/v1/add-user/123/"
+        reverse("authorization:grant-user-access-v2", kwargs=dict(pk=123))
+        == "/authorization/v2/grant-user-access/123/"
     )
     assert (
-        reverse("authorization:remove-user-v1", kwargs=dict(pk=123))
-        == "/authorization/v1/remove-user/123/"
+        reverse("authorization:revoke-user-access-v2", kwargs=dict(pk=123))
+        == "/authorization/v2/revoke-user-access/123/"
     )
     assert (
-        reverse("authorization:add-organization-v1", kwargs=dict(pk=123))
-        == "/authorization/v1/add-organization/123/"
+        reverse("authorization:grant-organization-access-v2", kwargs=dict(pk=123))
+        == "/authorization/v2/grant-organization-access/123/"
     )
     assert (
-        reverse("authorization:remove-organization-v1", kwargs=dict(pk=123))
-        == "/authorization/v1/remove-organization/123/"
+        reverse("authorization:revoke-organization-access-v2", kwargs=dict(pk=123))
+        == "/authorization/v2/revoke-organization-access/123/"
     )