+++ topobank/authorization/serializers.py
@@ -22,7 +22,7 @@ class UserPermissionSerializer(serializers.ModelSerializer):
     is_current_user = serializers.SerializerMethodField()
 
     def get_is_current_user(self, obj: UserPermission) -> bool:
-        return self.context['request'].user == obj.user
+        return self.context["request"].user == obj.user
 
 
 class OrganizationPermissionSerializer(serializers.ModelSerializer):
@@ -52,34 +52,39 @@ class PermissionSetSerializer(serializers.ModelSerializer):
         {
             "type": "object",
             "properties": {
-                "add_user": {"type": "string"},
-                "remove_user": {"type": "string"},
-                "add_organization": {"type": "string"},
-                "remove_organization": {"type": "string"},
+                "grant_user_access": {"type": "string"},
+                "revoke_user_access": {"type": "string"},
+                "grant_organization_access": {"type": "string"},
+                "revoke_organization_access": {"type": "string"},
             },
-            "required": ["add_user", "remove_user", "add_organization", "remove_organization"],
+            "required": [
+                "grant_user_access",
+                "revoke_user_access",
+                "grant_organization_access",
+                "revoke_organization_access",
+            ],
         }
     )
     def get_api(self, obj: PermissionSet) -> dict:
         request = self.context["request"]
         return {
-            "add_user": reverse(
-                "authorization:add-user-v1",
+            "grant_user_access": reverse(
+                "authorization:grant-user-access-v2",
                 kwargs={"pk": obj.id},
                 request=request,
             ),
-            "remove_user": reverse(
-                "authorization:remove-user-v1",
+            "revoke_user_access": reverse(
+                "authorization:revoke-user-access-v2",
                 kwargs={"pk": obj.id},
                 request=request,
             ),
-            "add_organization": reverse(
-                "authorization:add-organization-v1",
+            "grant_organization_access": reverse(
+                "authorization:grant-organization-access-v2",
                 kwargs={"pk": obj.id},
                 request=request,
             ),
-            "remove_organization": reverse(
-                "authorization:remove-organization-v1",
+            "revoke_organization_access": reverse(
+                "authorization:revoke-organization-access-v2",
                 kwargs={"pk": obj.id},
                 request=request,
             ),