+++ topobank/authorization/urls.py
@@ -2,12 +2,14 @@ from django.conf import settings
 from django.urls import path
 from rest_framework.routers import DefaultRouter, SimpleRouter
 
-from . import views as v1
+from . import views as v2
 
 app_name = "authorization"
 
 router = DefaultRouter() if settings.DEBUG else SimpleRouter()
-router.register(r"v1/permission-set", v1.PermissionSetViewSet, basename="permission-set-v1")
+router.register(
+    r"v2/permission-set", v2.PermissionSetViewSet, basename="permission-set-v2"
+)
 
 urlpatterns = router.urls
 
@@ -16,23 +18,23 @@ urlpatterns += [
     # API routes
     #
     path(
-        "v1/add-user/<pk>/",
-        view=v1.add_user,
-        name="add-user-v1",
+        "v2/grant-user-access/<pk>/",
+        view=v2.grant_user,
+        name="grant-user-access-v2",
     ),
     path(
-        "v1/remove-user/<pk>/",
-        view=v1.remove_user,
-        name="remove-user-v1",
+        "v2/revoke-user-access/<pk>/",
+        view=v2.revoke_user,
+        name="revoke-user-access-v2",
     ),
     path(
-        "v1/add-organization/<pk>/",
-        view=v1.add_organization,
-        name="add-organization-v1",
+        "v2/grant-organization-access/<pk>/",
+        view=v2.grant_organization,
+        name="grant-organization-access-v2",
     ),
     path(
-        "v1/remove-organization/<pk>/",
-        view=v1.remove_organization,
-        name="remove-organization-v1",
+        "v2/revoke-organization-access/<pk>/",
+        view=v2.revoke_organization,
+        name="revoke-organization-access-v2",
     ),
 ]