+++ topobank/settings/production.py
@@ -12,32 +12,6 @@ ALLOWED_HOSTS = env.list("DJANGO_ALLOWED_HOSTS", default=["example.com"])
 # ------------------------------------------------------------------------------
 DATABASES["default"]["CONN_MAX_AGE"] = env.int("CONN_MAX_AGE", default=60)  # noqa F405
 
-# CACHES
-# ------------------------------------------------------------------------------
-CACHES = {
-    "default": {
-        "BACKEND": env.str(
-            "DJANGO_DEFAULT_CACHE_BACKEND", default="django_redis.cache.RedisCache"
-        ),
-        "LOCATION": env.str(
-            "DJANGO_DEFAULT_CACHE_LOCATION", default="redis://redis:6379/0"
-        ),
-        "OPTIONS": {
-            "CLIENT_CLASS": "django_redis.client.DefaultClient",
-        },
-    }
-    # 'default': {
-    #     'BACKEND': 'django_redis.cache.RedisCache',
-    #     'LOCATION': env('REDIS_URL'),
-    #     'OPTIONS': {
-    #         'CLIENT_CLASS': 'django_redis.client.DefaultClient',
-    #         # Mimicing memcache behavior.
-    #         # http://niwinz.github.io/django-redis/latest/#_memcached_exceptions_behavior
-    #         'IGNORE_EXCEPTIONS': True,
-    #     }
-    # }
-}
-
 # SECURITY
 # ------------------------------------------------------------------------------
 # https://docs.djangoproject.com/en/dev/ref/settings/#secure-proxy-ssl-header